Move query string helpers to DataStore/helpers
