+++ src/RootStore/DataStore/helpers.js
@@ -1,3 +1,4 @@
+import qs from "qs";
 import toInteger from "lodash/fp/toInteger";
 import { convertFromStringToUnflattenedMatrix } from "../../api/metrics/optimizedFormatHelpers";
 import { parseResponseByFileFormat } from "../../api/metrics/fileParser";
@@ -26,4 +27,23 @@ export function processResponseData(data, file, eagerExpand = true) {
   }
 
   return metricFile;
+}
+
+/**
+ *
+ * @param {Object} filters - The filter values used to construct the query string to request metric data
+ * @param {string} filters.chargeCategory - A charge category or "All"
+ * @param {Array} filters.district - District IDs or "All"
+ * @param {string} filters.metricPeriodMonths - The number of months in the time period
+ * @param {string} filters.supervisionType - Supervision Type or "All"
+ * @param {string} filters.supervisionLevel - Supervision level or "All"
+ * @param {string} filters.reportedViolations - Number of reported violations or "All"
+ * @param {string} filters.violationType - Violation type or "All"
+ */
+export function getQueryStringFromFilters(filters = {}) {
+  return qs.stringify(filters, {
+    encode: false,
+    addQueryPrefix: true,
+    filter: (_, value) => (value !== "" ? value : undefined),
+  });
 }
