+++ src/RootStore/__tests__/helpers.test.js
@@ -1,5 +1,5 @@
 // Recidiviz - a data platform for criminal justice reform
-// Copyright (C) 2019 Recidiviz, Inc.
+// Copyright (C) 2021 Recidiviz, Inc.
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -15,7 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import { getQueryStringFromFilters } from "../urlHelpers";
+import { getQueryStringFromFilters } from "../DataStore/helpers";
 
 describe("getQueryStringFromFilters", () => {
   let filters;
