--- src/api/metrics/urlHelpers.js
@@ -1,38 +1,0 @@
-// Recidiviz - a data platform for criminal justice reform
-// Copyright (C) 2019 Recidiviz, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-// =============================================================================
-import qs from "qs";
-
-/**
- *
- * @param {Object} filters - The filter values used to construct the query string to request metric data
- * @param {string} filters.chargeCategory - A charge category or "All"
- * @param {Array} filters.district - District IDs or "All"
- * @param {string} filters.metricPeriodMonths - The number of months in the time period
- * @param {string} filters.supervisionType - Supervision Type or "All"
- * @param {string} filters.supervisionLevel - Supervision level or "All"
- * @param {string} filters.reportedViolations - Number of reported violations or "All"
- * @param {string} filters.violationType - Violation type or "All"
- */
-function getQueryStringFromFilters(filters = {}) {
-  return qs.stringify(filters, {
-    encode: false,
-    addQueryPrefix: true,
-    filter: (_, value) => (value !== "" ? value : undefined),
-  });
-}
-
-export { getQueryStringFromFilters };
